requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose
python-multipart
